Move Transformer to separate file
